+++ src/main/java/cn/wasu/community/community/controller/QuestionController.java
@@ -14,7 +14,7 @@ public class QuestionController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id")Integer id, Model model){
+    public String question(@PathVariable(name = "id")Long id, Model model){
         QuestionDTO questionDTO=questionService.getById(id);
         //累计阅读数
         questionService.incView(id);
