+++ src/main/java/cn/wasu/community/community/service/QuestionService.java
@@ -7,11 +7,11 @@ import cn.wasu.community.community.model.Question;
 public interface QuestionService {
     PaginationDTO list(Integer page, Integer size);
 
-    PaginationDTO list(Integer id, Integer page, Integer size);
+    PaginationDTO list(Long id, Integer page, Integer size);
 
-    QuestionDTO getById(Integer id);
+    QuestionDTO getById(Long id);
 
     void createOrUpdate(Question question);
 
-    void incView(Integer id);
+    void incView(Long id);
 }
